requirements.txt
selenium
webdriver-manager
fake-headers
pydantic
langchain-google-genai
openai
requests
python-dotenv
undetected-chromedriver
selenium-stealth
pytest
pytest-asyncio
